Split diagnostics explain tests by family
